Cleanup lang code.
