+++ crates/lang-node/src/pnpm.rs
@@ -1,14 +1,14 @@
+use crate::PNPM;
 use cached::proc_macro::cached;
 use moon_error::{map_io_to_fs_error, MoonError};
-use moon_lang::config_cache;
-use moon_lang::LockfileDependencyVersions;
+use moon_lang::{config_cache, LockfileDependencyVersions};
 use serde::{Deserialize, Serialize};
 use serde_yaml::Value;
 use std::collections::HashMap;
 use std::fs;
 use std::path::{Path, PathBuf};
 
-config_cache!(PnpmLock, "pnpm-lock.yaml", load_lockfile, write_lockfile);
+config_cache!(PnpmLock, PNPM.lock_filename, load_lockfile);
 
 type DependencyMap = HashMap<String, Value>;
 
@@ -54,10 +54,6 @@ fn load_lockfile<P: AsRef<Path>>(path: P) -> Result<PnpmLock, MoonError> {
     .map_err(|e| MoonError::Yaml(path.to_path_buf(), e))?;
 
     Ok(lockfile)
-}
-
-fn write_lockfile(_path: &Path, _lockfile: &PnpmLock) -> Result<(), MoonError> {
-    Ok(()) // Do nothing
 }
 
 #[cached(result)]
